cryptos-cool/requirements.txt
astroid==2.13.2
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
dill==0.3.6
h11==0.14.0
idna==3.4
isort==5.11.4
lazy-object-proxy==1.9.0
mccabe==0.7.0
platformdirs==2.6.2
pylint==2.15.10
python-dotenv==0.21.0
requests==2.28.1
tomlkit==0.11.6
typing_extensions==4.4.0
urllib3==1.26.13
uvicorn==0.20.0
wrapt==1.14.1
